Various improvements
Added support to Tokenizer and Parser classes to get strings
of delimiters between tokens.
Changed the way of generating time strings in Log to make it
more portable.
Entities has been updated to use integer values for positions
and sizes, in this way we prevent posible glitches rendering
to non integer positions.
Added the parameter OutlineQuality to TextEntity to control
the outline drawing quality.

+++ src/core/bmke/animations/move_in_animation.cpp
@@ -39,7 +39,7 @@ MoveInAnimation::~MoveInAnimation(void)
 
 void MoveInAnimation::init(Entity* entity)
 {
-	sf::Vector2f size;
+	sf::Vector2i size;
 
 	Animation::init(entity);
