+++ src/core/bmke/animations/move_out_animation.cpp
@@ -39,7 +39,7 @@ MoveOutAnimation::~MoveOutAnimation(void)
 
 void MoveOutAnimation::init(Entity* entity)
 {
-	sf::Vector2f size;
+	sf::Vector2i size;
 
 	Animation::init(entity);
